requirements.txt
azure-common==1.1.26
azure-core==1.11.0
azure-functions==1.6.0
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.8.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0
azure-storage-file-share==12.4.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
isodate==0.6.0
msrest==0.6.21
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.26.3
